feat: refactor vite plugin
