+++ src/vite/mod.ts
@@ -19,6 +19,6 @@ export type {
   LimetteRouteManifestEntry,
 } from './manifest.ts';
 export { limette } from './plugin.ts';
-export { clientEntryDevPath } from './plugin.ts';
+export { clientEntryDevPath } from './client-entry.ts';
 export type { LimetteVitePluginOptions } from './plugin.ts';
 export type { IslandImport } from './islands.ts';
